pyproject.toml
[project]
name = "soniox-transcriber"
version = "0.1.0"
description = "Live audio transcription app using Soniox API"
dependencies = [
    "websockets>=12.0",
    "pyaudio>=0.2.13",
    "python-dotenv>=1.0.0",
    "pyautogui>=0.9.54",
    "pynput>=1.7.6",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
]
requires-python = ">=3.8"

[project.scripts]
soniox-transcriber = "soniox_transcriber.__main__:main"
soniox-dictate = "soniox_transcriber.dictation:main"
soniox-vapi-server = "soniox_transcriber.vapi_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/soniox_transcriber"]